\name{spatstat.gui-internal} 
\title{Internal functions of spatstat.gui package}
\alias{do.istat}
\description{
  Internal functions.
}
\usage{
do.istat(panel)
}
\details{
  These internal functions of the \pkg{spatstat.gui} package
  are not usually called
  directly by the user. Their names and capabilities may change
  without warning from one version of \pkg{spatstat.gui} to the next.
}
\keyword{internal}
