\name{spatstat.gui-package}
\alias{spatstat.gui-package}
\alias{spatstat.gui}
\docType{package}
\title{
  Interactive Graphics for the 'spatstat' Package
}
\description{
  Extension of the spatstat package, containing
  interactive graphics.
}
\details{
  The main functions are
  \tabular{ll}{
    \code{\link{iplot}} \tab Interactive plot of spatial dataset \cr
    \code{\link{istat}} \tab Interactive statistical summaries 
  }
}
\author{
  \spatstatAuthors.
}
\references{
  Baddeley, A., Rubak, E. and Turner, R. (2015)
  \emph{Spatial Point Patterns: Methodology and Applications with R}.
  Chapman and Hall/CRC Press.
}
\keyword{package}
